+++ .eslintrc.json
@@ -17,6 +17,8 @@
     "import/extensions": "off",
     "import/no-unresolved": "off",
     "class-methods-use-this": "off",
+    "no-useless-constructor": "off",
+    "no-empty-function": "off",
     "@typescript-eslint/no-unused-vars": ["error"]
   }
 }
