+++ src/config/app.ts
@@ -2,8 +2,8 @@ import express from 'express';
 import cors from 'cors';
 import morgan from 'morgan';
 import SwaggerUi from 'swagger-ui-express';
-import swagger from './docs/swagger';
-import routerIndex from './routes/index.Routes';
+import swagger from './swagger';
+import routerIndex from '../shared/index.Routes';
 
 const app = express();
