--- src/controllers/suma.ctrl.ts
@@ -1,4 +1,0 @@
-function sum(a: number, b: number): number {
-  return a + b;
-}
-export default sum;
