+++ test/suma.spec.ts
@@ -1,4 +1,4 @@
-import sum from '../src/controllers/suma.ctrl';
+import app from '../src/config/app';
 
 describe('sum module', () => {
   test('adds 1 + 2 to equal 3', () => {
